refactor(core-weave): extract Knop support preservation renderer
- move carried KnopSourceRegistry and ReferenceCatalog preservation helpers into knop_support_renderers
- keep weave.ts as the planner dispatcher and public façade
- preserve generated Turtle behavior while reducing weave.ts size
- update core weave module overview
